Create a single cleaned addressbase file
This is because postgres COPY performs *really* badly when there is a
lot of existing data in a table. Splitting up the files meant each file
was taking much longer than the previous one.

At least this way, postgres will be able to figure out how much work
it has to do upfront